Inflector.Inflector — Module

Handles the functionality for applying various gramatical rules.

Inflector.IRREGULAR_NOUNS — Constant

Vector of nouns with irregular forms of singular and/or plural. Each tuple contains pairs of singular and plural forms.

Inflector.from_dashes — Function

from_dashes(word::String) :: String

Generates SnakeCase form of word from dash_case.

Inflector.from_underscores — Function

from_underscores(word::String) :: String

Generates SnakeCase form of word from underscore_case.

Inflector.irregular — Method

irregular(word::String) :: Union{Nothing,Tuple{String,String}}

Wether or not word has an irregular singular or plural form.

Inflector.irregulars — Method

irregulars() :: Vector{Tuple{String,String}}

Returns a vector of words with irregular singular or plural forms.

Inflector.is_irregular — Method

function is_irregular(word::String) :: Bool

Whether or not word has a singular or plural irregular form.

Inflector.is_plural — Method

is_plural(word::String) :: Bool

Returns wether or not word is a plural.

Inflector.is_singular — Method

is_singular(word::String) :: Bool

Returns wether or not word is a singular.

Inflector.to_plural — Method

to_plural(word::String; is_irregular::Bool = false) :: Union{Nothing,String}

Returns the plural form of word.

Inflector.to_plural_irregular — Method

to_plural_irregular(word::String) :: Union{Nothing,String}

Returns the plural form of the irregular word word.

Inflector.to_singular — Method

to_singular(word::String; is_irregular::Bool = false) :: Union{Nothing,String}

Returns the singural form of word.

Inflector.to_singular_irregular — Method

to_singular_irregular(word::String) :: Union{Nothing,String}

Returns the singular form of the irregular word word.